% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{add_to_phyloseq_ordination}
\alias{add_to_phyloseq_ordination}
\title{Add new points to an ordination object created by phyloseq}
\usage{
add_to_phyloseq_ordination(ps_old, ps_old_and_new, distance)
}
\arguments{
\item{ps_old}{A phyloseq object containing the original samples used for phyloseq::ordinate.}

\item{ps_old_and_new}{A phyloseq object containing both the old samples (those used for phyloseq::ordinate) and the new samples to be added to the embedding diagram.}

\item{distance}{The same argument that was passed to phyloseq::ordinate}
}
\description{
Add new points to an ordination object created by phyloseq
}
