% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{embed_new_points}
\alias{embed_new_points}
\title{Embed new points in an MDS diagram}
\usage{
embed_new_points(
  mds_matrices,
  new_points = NULL,
  dist_fn = NULL,
  old_new_dist_mat = NULL
)
}
\arguments{
\item{mds_matrices}{The output from make_mds_matrices.}

\item{new_points}{A matrix, each row corresponding to a new sample, each column corresponding to a variable.}

\item{dist_fn}{A function taking a matrix as its argument, returns distances between the rows.}

\item{old_new_dist_mat}{A matrix with n_old rows and n_new columns containing distances between the old and new points. If this argument is non-null, these distances will be used instead of computing the distances using dist_fn.}
}
\value{
A matrix containing the embedding locations of the new points. Rows correspond to new samples, columns correspond to embedding dimensions.
}
\description{
Embed new points in an MDS diagram
}
