% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdessence-functions.R
\name{get_distances}
\alias{get_distances}
\title{Get distances between a set of old points and a set of new points}
\usage{
get_distances(old_points, new_points, dist_fn)
}
\arguments{
\item{old_points}{n_old x n_variables matrix with rows containing the old points.}

\item{new_points}{n_new x n_variables matrix with rows containing the new point.}

\item{dist_fn}{A function taking as input a matrix, returning the distances between the rows of the matrix.}
}
\value{
A matrix with n_old rows and n_new columns containing the
    distance between each pair of old and new points.
}
\description{
Get distances between a set of old points and a set of new points
}
